✨ feat(events): add new events for player interactions and state changes
- introduce AfkStateChangeEvent for tracking player's AFK status changes
- add ArchaeologySpecialDropEvent for special item drops during archaeology
- implement ArmorStandBreakWithAxeEvent for axe interactions with armor stands
- create CaptchaFailEvent and CaptchaSuccessEvent for CAPTCHA verification results
- add ChatAiModerateEvent for AI moderation of chat messages
- implement ChatDirectMessageEvent for private messaging between players
- create ChatIgnoreEvent and ChatUnIgnoreEvent for managing player ignore lists
- add Clan events for creation, deletion, promotion, demotion, and invite count changes
- introduce Player events for playtime milestones and payment transactions
- add various item-related events including ShopBuyItemEvent and ShopSellItemEvent

+++ surf-api-paper/surf-api-paper/src/main/kotlin/dev/slne/surf/api/paper/event/protect/SellRegionEvent.kt
@@ -2,9 +2,20 @@ package dev.slne.surf.api.paper.event.protect
 
 import dev.slne.surf.api.core.event.SurfSyncEvent
 import org.bukkit.entity.Player
+import java.math.BigDecimal
 
+/**
+ * Fired when a player sells a protected region.
+ *
+ * This synchronous event is triggered when a player sells off a protected region,
+ * receiving monetary compensation for it.
+ *
+ * @property player The player who sold the region
+ * @property size The size of the sold region
+ * @property receivedMoney The amount of money received from the sale
+ */
 data class SellRegionEvent(
     val player: Player,
     val size: Int,
-    val receivedMoney: Double
+    val receivedMoney: BigDecimal
 ) : SurfSyncEvent()
